requirements.txt
torch
torchvision
transformers==4.33
decord
editdistance==0.8.1
einops==0.7.0
evaluate==0.4.1
fairscale
fastapi==0.110.0
ftfy==6.1.3
google_api_python_client==2.119.0
gradio
# horovod==0.28.1  # 注释掉，因为构建失败且非必需
huggingface_hub
iopath==0.1.10
matplotlib
omegaconf==2.3.0
openai
opencv_python
opendatasets==0.1.22
packaging==24.0
pandas==2.2.1
peft==0.6.2
Pillow
pycocoevalcap
pycocotools==2.0.7
regex
scipy
seaborn
sentencepiece
shortuuid
scikit-image
timm
tqdm
uvicorn
webdataset
accelerate
bitsandbytes